Floats to doubles (#5)
* change floats to doubles

* add test scenes

* change fmaxf to fmax

+++ miniRT.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   miniRT.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ralba-ji <ralba-ji@student.42madrid.com    +#+  +:+       +#+        */
+/*   By: isastre- <isastre-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/01 14:33:54 by ralba-ji          #+#    #+#             */
-/*   Updated: 2026/03/31 14:26:25 by ralba-ji         ###   ########.fr       */
+/*   Updated: 2026/03/31 17:43:50 by isastre-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -228,13 +228,13 @@ void		manage_hooks(t_mlxinfo *window);
 //     render
 void		render(t_miniRT *scene);
 t_ray		create_ray(int x, int y, t_miniRT *scene);
-float		intersect(t_ray ray, t_list *scene, t_list **obj);
+double		intersect(t_ray ray, t_list *scene, t_list **obj);
 t_color		get_color(t_list lst);
 int			color_to_int(t_color c);
 t_color		color_sum(t_color a, t_color b);
 t_color		fminf_color(t_color color1, t_color color2, int max);
-t_color		apply_light(t_color obj, float intensity, t_color light_color);
+t_color		apply_light(t_color obj, double intensity, t_color light_color);
 t_3dvector	get_normal(t_list *hit, t_3dvector hit_point);
-bool		intersect_ray_cylinder(t_ray *ray, t_cylinder *cyl, float *t);
+bool		intersect_ray_cylinder(t_ray *ray, t_cylinder *cyl, double *t);
 
 #endif
